requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-oauth-toolkit>=2.3.0
django-filter>=23.1
mock>=5.0.0
Pillow>=9.0.0
django-guardian>=2.4.0
celery[redis]>=5.2.0
cryptography>=41.0.0
twilio>=8.0.0
requests>=2.31.0
coverage>=7.2.7
python-dotenv>=1.0.0
